backend/pyproject.toml
[project]
name = "triequest-backend"
version = "0.1.0"
description = "TrieQuest backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16.4",
  "email-validator>=2.2.0",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "PyMySQL>=1.1.1",
  "pydantic-settings>=2.10.1",
  "pwdlib[argon2]>=0.2.1",
  "PyJWT>=2.10.1",
  "sqlalchemy>=2.0.43",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "ruff>=0.15.6",
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"
